// src/types/Character.ts
export type Character = {
  _id: number;
  url: string;
  name: string;
  sourceUrl: string;
  imageUrl: string;
  films: string[];
  shortFilms: string[];
  tvShows: string[];
  videoGames: string[];
  alignment: string;
  parkAttractions: string[];
  allies: string[];
  enemies: string[];
  updatedAt: string;
};